refactor: restructure memory store modules into memory_stores/ package
- base/memory.py split: BaseMemoryStore moved to base/memory_store.py;
  BaseMemory stays until Recency/Similarity/Reflective are refactored
- New memory_stores/ package: json.py, qdrant/store.py, qdrant/utils.py
- Deleted memory/json_store.py, memory/qdrant_store.py, memory/qdrant_utils.py
- All imports and test patch paths updated throughout
- Add .claude/settings.json to .gitignore
