+++ src/llm_agents_from_scratch/memory/__init__.py
@@ -1,8 +1,11 @@
 """Memory implementations."""
 
-from .json_store import JSONMemoryStore
+from llm_agents_from_scratch.memory_stores import (
+    JSONMemoryStore,
+    QdrantMemoryStore,
+)
+
 from .memory import Memory, MetadataFn
-from .qdrant_store import QdrantMemoryStore
 from .recency import RecencyMemory
 from .reflective import ReflectiveMemory
 from .similarity import SimilarityMemory
