+++ src/llm_agents_from_scratch/memory/memory.py
@@ -4,7 +4,7 @@ import asyncio
 import inspect
 from collections.abc import Awaitable, Callable
 
-from llm_agents_from_scratch.base.memory import BaseMemoryStore
+from llm_agents_from_scratch.base.memory_store import BaseMemoryStore
 from llm_agents_from_scratch.data_structures import Task
 from llm_agents_from_scratch.data_structures.memory import Episode
