+++ src/llm_agents_from_scratch/memory/recency.py
@@ -1,6 +1,7 @@
 """Recency-based episodic memory."""
 
-from llm_agents_from_scratch.base.memory import BaseMemory, BaseMemoryStore
+from llm_agents_from_scratch.base.memory import BaseMemory
+from llm_agents_from_scratch.base.memory_store import BaseMemoryStore
 from llm_agents_from_scratch.data_structures import Task
 from llm_agents_from_scratch.data_structures.memory import Episode
