+++ src/llm_agents_from_scratch/memory/similarity.py
@@ -2,7 +2,8 @@
 
 from typing import Any
 
-from llm_agents_from_scratch.base.memory import BaseMemory, BaseMemoryStore
+from llm_agents_from_scratch.base.memory import BaseMemory
+from llm_agents_from_scratch.base.memory_store import BaseMemoryStore
 from llm_agents_from_scratch.data_structures import Task
 from llm_agents_from_scratch.data_structures.memory import Episode
