+++ src/llm_agents_from_scratch/memory_stores/qdrant/__init__.py
@@ -1,0 +1,5 @@
+"""Qdrant-backed episodic memory store."""
+
+from .store import QdrantMemoryStore
+
+__all__ = ["QdrantMemoryStore"]
