+++ tests/memory/test_base.py
@@ -1,4 +1,5 @@
-from llm_agents_from_scratch.base.memory import BaseMemory, BaseMemoryStore
+from llm_agents_from_scratch.base.memory import BaseMemory
+from llm_agents_from_scratch.base.memory_store import BaseMemoryStore
 
 
 def test_base_memory_abstract_methods() -> None:
